1.Add scan module, drivers can scan packages via head office App and EasyDelivery at the same time.

+++ app/src/main/java/com/hf/easydelivery/view/LoginDialog.java
@@ -1,6 +1,6 @@
 package com.hf.easydelivery.view;
 
-import android.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.view.LayoutInflater;
@@ -57,7 +57,7 @@ public class LoginDialog extends AlertDialog {
                             ResourceMgr.getInstance().getLoginInfo().loginName = name;
 
                             if (isNumeric(name)) {
-                                ResourceMgr.getInstance().getLoginInfo().loginId = Short.valueOf(name);
+                                ResourceMgr.getInstance().getLoginInfo().loginId = Integer.valueOf(name);
 
                                 ICourierService courierService = ResourceMgr.getInstance().getCourierService();
                                 assert courierService != null;
